add functions to set content-type based on file extension

+++ srcs/response/include/server_config.hpp
@@ -18,11 +18,11 @@ class ServerConfig {
   */
   LocationConfig* SelectLocationConfig(const std::string& uri);
 
-  int port_; // The highest TCP port number is 65535
+  size_t port_; // The highest TCP port number is 65535
   std::string host_;
   std::vector<std::string> server_names_;
   std::string error_page_path_;
-  int client_max_body_size_; // @fyuta: ConfigParserの方で数字に読み替えてもらえるとありがたいです。intで収まり切るかどうかは要検討
+  size_t client_max_body_size_; // @fyuta: ConfigParserの方で数字に読み替えてもらえるとありがたいです。intで収まり切るかどうかは要検討
   std::vector<LocationConfig> vec_location_config_;
 };
